@@ -26,7 +26,9 @@
         <p>
           I write and speak, some of which you can find below. I'm also
           available for work: contact me at
-          <a href="mailto:callum@macr.ae">callum@macr.ae</a>.
+          <a
+            href="mailto:callum@macr.ae"
+          >callum@macr.ae</a>.
         </p>
         <a class="down-arrow" role="presentation" @click="handleScrollClick">
           <FontAwesomeIcon
@@ -37,27 +39,15 @@
           />
         </a>
         <div class="social-links">
-          <a
-            href="https://twitter.com/callumacrae"
-            target="_blank"
-            rel="noopener"
-          >
+          <a href="https://twitter.com/callumacrae" target="_blank" rel="noopener">
             <span class="sr-only">Twitter</span>
             <FontAwesomeIcon :icon="['fab', 'twitter']" />
           </a>
-          <a
-            href="https://github.com/callumacrae"
-            target="_blank"
-            rel="noopener"
-          >
+          <a href="https://github.com/callumacrae" target="_blank" rel="noopener">
             <span class="sr-only">Github</span>
             <FontAwesomeIcon :icon="['fab', 'github']" />
           </a>
-          <a
-            href="https://codepen.io/callumacrae"
-            target="_blank"
-            rel="noopener"
-          >
+          <a href="https://codepen.io/callumacrae" target="_blank" rel="noopener">
             <span class="sr-only">Codepen</span>
             <FontAwesomeIcon :icon="['fab', 'codepen']" />
           </a>
@@ -73,16 +63,10 @@
           clipPath: rightTitleClipPath,
           '-webkit-clip-path': rightTitleClipPath
         }"
-      >
-        Callum Macrae
-      </h1>
+      >Callum Macrae</h1>
 
       <div class="right__content">
-        <a
-          href="https://twitter.com/callumacrae"
-          target="_blank"
-          rel="noopener"
-        >
+        <a href="https://twitter.com/callumacrae" target="_blank" rel="noopener">
           <span class="sr-only">Twitter</span>
           <FontAwesomeIcon :icon="['fab', 'twitter']" />
         </a>
